--HG-- rename : UModSys/src/module/media/stdlib/umodsys.mod.media.stdlib.t.binarchiveframe.h => UModSys/src/module/media/stdlib/umodsys.mod.media.stdlib.Tbinarchiveframe.h rename : UModSys/src/module/media/stdlib/umodsys.mod.media.stdlib.t.libbincache.h => UModSys/src/module/media/stdlib/umodsys.mod.media.stdlib.Tlibbincache.h rename : UModSys/src/module/media/stdlib/umodsys.mod.media.stdlib.t.libbintree.h => UModSys/src/module/media/stdlib/umodsys.mod.media.stdlib.Tlibbintree.h rename : UModSys/src/module/media/stdlib/umodsys.mod.media.stdlib.t.liblayered.h => UModSys/src/module/media/stdlib/umodsys.mod.media.stdlib.Tliblayered.h rename : UModSys/src/module/media/stdlib/umodsys.mod.media.stdlib.t.libobjcache.h => UModSys/src/module/media/stdlib/umodsys.mod.media.stdlib.Tlibobjcache.h rename : UModSys/src/module/media/stdlib/umodsys.mod.media.stdlib.t.libobjfilter.h => UModSys/src/module/media/stdlib/umodsys.mod.media.stdlib.Tlibobjfilter.h rename : UModSys/src/module/media/stdlib/umodsys.mod.media.stdlib.t.utils.h => UModSys/src/module/media/stdlib/umodsys.mod.media.stdlib.Tutilities.h rename : UModSys/src/module/media/zlib/umodsys.mod.media.zlib.arcframe.h => UModSys/src/module/media/zlib/umodsys.mod.media.zlib.Tarcframe.h extra : convert_revision : svn%3Ac7b0911a-351f-11e0-b982-4926ce03a894%40551

+++ include/umodsys/tl/string/su_comparer.h
@@ -50,7 +50,7 @@ struct TComparerBinary : public core::Void {
 // Comparer hash
 
 struct SComparerHash {
-  int hash;
+  size_t hash;
   //
   inline SComparerHash(void) : hash(0) {}
   inline SComparerHash(core::Void* p) {}
@@ -83,13 +83,13 @@ struct TComparerBinaryHash : public SComparerHash {
     return myl==rl && su::scmp(mys, r, myl)==0;
   }
   //
-  inline int cmp3(const CharT* mys, size_t myl, const CharT* r, size_t rl, int rh) const {
+  inline int cmp3(const CharT* mys, size_t myl, const CharT* r, size_t rl, size_t rh) const {
     int rv = core::scalar_compare(myl, rl);
     if(rv) return rv;
     rv = core::scalar_compare(hash, rh);
     return rv ? rv : su::scmp(mys, r, myl);
   }
-  inline bool eq3(const CharT* mys, size_t myl, const CharT* r, size_t rl, int rh) const {
+  inline bool eq3(const CharT* mys, size_t myl, const CharT* r, size_t rl, size_t rh) const {
     return myl==rl && hash==rh && su::scmp(mys, r, myl)==0;
   }
   inline int cmp3(const CharT* mys, size_t myl, const CharT* r, size_t rl, const SComparerHash& rh) const {
@@ -130,13 +130,13 @@ struct TComparerUtfNocaseHash : public SComparerHash {
     return myl==rl && su::utf_cmp_nocase(mys, r, myl)==0;
   }
   //
-  inline int cmp3(const CharT* mys, size_t myl, const CharT* r, size_t rl, int rh) const {
+  inline int cmp3(const CharT* mys, size_t myl, const CharT* r, size_t rl, size_t rh) const {
     int rv = core::scalar_compare(myl, rl);
     if(rv) return rv;
     rv = core::scalar_compare(hash, rh);
     return rv ? rv : su::utf_cmp_nocase(mys, r, myl);
   }
-  inline bool eq3(const CharT* mys, size_t myl, const CharT* r, size_t rl, int rh) const {
+  inline bool eq3(const CharT* mys, size_t myl, const CharT* r, size_t rl, size_t rh) const {
     return myl==rl && hash==rh && su::scmp(mys, r, myl)==0;
   }
   inline int cmp3(const CharT* mys, size_t myl, const CharT* r, size_t rl, const SComparerHash& rh) const {
